… dist/synaudio.js
… package.json