## Features
* Slight performance gain by only calculating comparison clip standard deviation once
  * SIMD tests from 24 seconds to 22 seconds
  * Scalar tests from 76 seconds to 67 seconds
* Better validation on inputs

## Fixes
* Use double precision for mean sum to avoid potential overflow
* Shrink `correlationSampleSize` if either clip is smaller to prevent segfault